requirements.txt
aiohttp>=3.8.0
aiofiles>=23.0.0
pydantic>=2.0.0
jinja2>=3.1.0
pyyaml>=6.0
requests>=2.28.0
asyncio-mqtt>=0.16.0
fastapi>=0.100.0
uvicorn>=0.23.0
# 通义千问API通过HTTP直接调用，无需额外SDK